lib: support rotating output files to limit file size
The file that stores the profiling events is not bounded to any limit in terms
of file size or number of lines.

In some configurations like embedded systems, the storage space is limited. Let's
add a new parameter to start() function for capping the file size to
a maximum. When this parameter is given, two rotating files are created
to hold the more recent events and never exceed the max size.

Fix issue #13