libs/eventsfile: fix race condition when rotating log file
The condition to know if file rotation should take place can be accessed
in a concurrent manner by 2 or more threads. One thread can enter the if
block and open a new file and close the current one while the other thread
still tries to write to the current logfile that has been closedl!
This leads to segfault.

Let's move the mutex before the block to prevent this race condition.

Signed-off-by: Cedric CHEDALEUX <cedric.chedaleux@orange.com>